Remaked some commands

+++ WafclastRPG/Extensions/RandomExtension.cs
@@ -6,6 +6,7 @@ namespace WafclastRPG.Extensions
     {
         public static bool Chance(this Random rd, double chance)
             => rd.NextDouble() < chance;
+
         public static int Sortear(this Random rd, int max)
             => rd.Next(0, max);
 
@@ -15,7 +16,7 @@ namespace WafclastRPG.Extensions
         public static double Sortear(this Random rd, double min, double max)
             => rd.NextDouble() * (max - min) + min;
 
-        public static decimal Sortear(this Random rd, decimal max)
-            => (decimal)(rd.NextDouble() * (double)max);
+        public static double Sortear(this Random rd, double max)
+            => (double)(rd.NextDouble() * (double)max);
     }
 }
